requirements.txt
# Database and Vector Support
psycopg2-binary>=2.9.10
pgvector>=0.2.4
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# AI/ML
openai>=1.35.0
numpy>=2.1.0
pandas>=2.2.3

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3

# HTTP and Scraping
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0